% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_affair_descriptors}
\alias{get_affair_descriptors}
\title{Retrieve affair descriptors}
\usage{
get_affair_descriptors(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of descriptors.
\item \code{updated} date of last update.
\item \code{code} code of descriptor.
\item \code{name} language-specific descriptor.
}
}
\description{
\code{get_affair_descriptors} retrieves all descriptors used to classify the content of affairs of the Swiss Parliament.
}
\examples{
\donttest{
get_affair_descriptors(lang = "fr")
}
}
