% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affairs.R
\name{get_affair_details}
\alias{get_affair_details}
\title{Retrieve data on political affairs}
\usage{
get_affair_details(affair_id, lang = "de", silent = F)
}
\arguments{
\item{affair_id}{numeric or character vector of affair IDs to query.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 17 columns, including
\itemize{
\item \code{affairType} a list with data on the affair type.
\item \code{author} a list with data on the author of the affair.
\item \code{draft} a list with data on the consultation of the affair.
\item \code{texts} a list with associated texts.
}
}
\description{
\code{get_affair_details} retrieves detailed data on political affairs of the Swiss Parliament provided by
    the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_affair_details(affair_id = c(20130468, 20123678))
}
}
