% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_affair_states}
\alias{get_affair_states}
\title{Retrieve affair states}
\usage{
get_affair_states(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of affair state.
\item \code{updated} date of last update.
\item \code{code} code of affair state.
\item \code{name} full language-specific description of affair type.
\item \code{sorting} sorting vector.
\item \code{comment} additional information on use of affair state.
\item \code{parent.id} additional ID.
}
}
\description{
\code{get_affair_states} retrieves all possible states of affairs of the Swiss Parliament.
}
\examples{
\donttest{
get_affair_states(lang = "fr")
}
}
