% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affairs.R
\name{get_affairs2}
\alias{get_affairs2}
\title{Retrieve data on political affairs}
\usage{
get_affairs2(year = NULL, lang = "de", silent = F)
}
\arguments{
\item{year}{numeric or character vector of years to query. If \code{is.null(year)} all available affairs are retrieved.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 17 columns, including:
\itemize{
\item \code{affairType} a list with data on the affair type.
\item \code{author} a list with data on the author of the affair.
\item \code{draft} a list with data on the consultation of the affair.
\item \code{texts} a list with associated texts.
}
}
\description{
\code{get_affairs2} is wrapper around \code{get_affairs} and enables the mass retrieval of data
    on political affairs of the Swiss Parliament by the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_affairs2(year = c(1999:2003))
}
}
