% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affairsummaries.R
\name{get_affairsummaries}
\alias{get_affairsummaries}
\title{Retrieve affair summaries}
\usage{
get_affairsummaries(affair_id, lang = "de", silent = F)
}
\arguments{
\item{affair_id}{numeric or character vector of affair IDs to query.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 7 columns, including \code{description}, \code{initialSituation} and \code{proceedings}.
}
\description{
\code{get_affairsummaries} retrieves summaries of political affairs provided by the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_affairsummaries(affair_id = c(20130468, 19950056))
}
}
