% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_cantons}
\alias{get_cantons}
\title{Retrieve basic cantonal information}
\usage{
get_cantons(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of canton.
\item \code{updated} date of last update.
\item \code{abbreviation} short form of canton.
\item \code{code} code of canton.
\item \code{name} full language-specific description of canton.
}
}
\description{
\code{get_cantons} retrieves basic information on cantons.
}
\examples{
\donttest{
get_cantons(lang = "fr")
}
}
