% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_committees}
\alias{get_committees}
\title{Retrieve basic committee information}
\usage{
get_committees(main = F, permanent = F, current = F, lang = "de",
  silent = F)
}
\arguments{
\item{main}{if \code{TRUE} only main committees are retrieved.}

\item{permanent}{if \code{TRUE} only permanent committees are retrieved.}

\item{current}{if \code{TRUE} only current committees are retrieved.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 17 columns, including:
\itemize{
\item \code{id} ID of committee.
\item \code{abbreviation} code of committee.
\item \code{from} date of first convocation.
\item \code{isActive} status of activity.
\item \code{name} full language-specific description of committee.
\item \code{to} date of dissolution.
}
}
\description{
\code{get_committees} retrieves basic information on all committees of the Swiss Parliament.
}
\examples{
\donttest{
get_committees(lang = "fr")
}
}
