% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_councillor_details.R
\name{get_councillor_details2}
\alias{get_councillor_details2}
\title{Retrieve details on all councillors}
\usage{
get_councillor_details2(current = FALSE, lang = "de", silent = F)
}
\arguments{
\item{current}{if \code{TRUE} only current councillors are retrieved.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 40 columns, including \code{canton}, \code{birthDate}, \code{committeeMemberships},
    \code{maritalStatus}, \code{numberOfChildren} and \code{militaryGrade}.
}
\description{
\code{get_councillor_details2} is wrapper around \code{get_councillor_details} and enables the mass retrieval of
    detailed data on councillors provided by the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_councillor_details(current = T)
}
}
