% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_councillors}
\alias{get_councillors}
\title{Retrieve current councillors}
\usage{
get_councillors(council = NULL, canton = NULL, lang = "de",
  silent = F)
}
\arguments{
\item{council}{specifies council. Use \code{council = "N"} for the National Council and \code{council = "S"} for
the Council of States to obtain council-specific results.}

\item{canton}{specifies canton. Use two-letter abbreviations to obtain canton-specific results.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 16 columns, including:
\itemize{
\item \code{id} ID of councillor.
\item \code{biographyUrl} link to official biography.
\item \code{canton} canton affiliation.
\item \code{council} council affiliation.
\item \code{faction} faction affiliation.
\item \code{party} party affiliation.
\item \code{homepage} url of personal homepage.
}
}
\description{
\code{get_councillors} retrieves the current councillors of the Swiss Parliament.
}
\examples{
\donttest{
get_councillors(lang = "fr")
}
}
