% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_councillors_historic}
\alias{get_councillors_historic}
\title{Retrieve all councillors}
\usage{
get_councillors_historic(date_of_death_from = NULL,
  date_of_death_to = NULL, faction = NULL,
  legislative_period_from = NULL, lang = "de", silent = F)
}
\arguments{
\item{date_of_death_from}{filters by date of death after a given date. Recommended date format \code{YYYY/MM/DD}.}

\item{date_of_death_to}{filters by date of death before a given date. Recommended date format \code{YYYY/MM/DD}.}

\item{faction}{specifies faction affiliation.}

\item{legislative_period_from}{specifies legislative period.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 28 columns, including:
\itemize{
\item \code{id} ID of councillor.
\item \code{birthDate} date of birth.
\item \code{dateOfDeath} date of death.
\item \code{membership.entryDate} Council specific terms of office, start.
\item \code{membership.leavingDate} Council specific terms of office, end.
}
}
\description{
\code{get_councillors_historic} retrieves all councillors in the history of the Swiss Parliament.
}
\examples{
\donttest{
get_councillors_historic(lang = "fr")
}
}
