% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_councils}
\alias{get_councils}
\title{Retrieve basic council information}
\usage{
get_councils(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of council.
\item \code{updated} date of last update.
\item \code{abbreviation} short form of council.
\item \code{code} code of council.
\item \code{name} language-specific name of council.
}
}
\description{
\code{get_councils} retrieves basic information on the councils of the Swiss Parliament.
}
\examples{
\donttest{
get_councils(lang = "fr")
}
}
