% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_departments_historic}
\alias{get_departments_historic}
\title{Retrieve historical information on the departments}
\usage{
get_departments_historic(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of department.
\item \code{updated} date of last update.
\item \code{abbreviation} short form of department.
\item \code{code} code of department.
\item \code{name} language-specific name of department.
}
}
\description{
\code{get_departments_historic} retrieves basic information on all departments in the history of the Swiss government.
}
\examples{
\donttest{
get_departments_historic(lang = "fr")
}
}
