% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_faction_members2}
\alias{get_faction_members2}
\title{Retrieve all faction memberships}
\usage{
get_faction_members2(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 22 columns, including \code{canton}, \code{council}, \code{party} and \code{factionFunction}.
}
\description{
\code{get_faction_members2} is wrapper around \code{get_faction_members} and enables the mass retrieval of
    data on faction memberships provided by the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_faction_members2()
}
}
