% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_factions_historic}
\alias{get_factions_historic}
\title{Retrieve historical information on the factions}
\usage{
get_factions_historic(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of faction.
\item \code{updated} date of last update.
\item \code{abbreviation} short form of faction.
\item \code{code} code of faction.
\item \code{from} first presence in Parliament.
\item \code{name} language-specific name of faction.
\item \code{shortName} language-specific short name of faction.
\item \code{to} last presence in Parliament.
}
}
\description{
\code{get_factions_historic} retrieves basic information on all factions in the history of the Swiss Parliament.
}
\examples{
\donttest{
get_factions_historic(lang = "fr")
}
}
