% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glimpse.R
\name{get_glimpse2}
\alias{get_glimpse2}
\title{Retrieve the first rows of an OpenParlData resource}
\usage{
get_glimpse2(table, rows = 20)
}
\arguments{
\item{table}{name of the OpenParlData resource to glimpse
into. For an overview of available endpoints use
\code{\link{get_tables2}()}.}

\item{rows}{number of records to download. Maximum is 1000.}
}
\value{
A tibble containing up to \code{rows} records. Column composition
  depends on the selected OpenParlData resource.
}
\description{
\code{get_glimpse2} retrieves the first rows of a resource provided by the
OpenParlData.ch REST API and allows a first insight into its data structure.
}
\examples{
\dontrun{
# Short excerpt of OpenParlData resource "persons"
get_glimpse2(table = "persons")
}
}
