% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_legislative_periods}
\alias{get_legislative_periods}
\title{Retrieve legislative periods}
\usage{
get_legislative_periods(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of legislative period.
\item \code{updated} date of last update.
\item \code{code} code of legislative period.
\item \code{from} start date of legislative period.
\item \code{name} full language-specific description of legislative period.
\item \code{to} end date of legislative period.
}
}
\description{
\code{get_legislative_periods} retrieves basic information on all available legislative periods.
}
\examples{
\donttest{
get_legislative_periods(lang = "fr")
}
}
