% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{Retrieve overview data}
\usage{
get_overview(ws_page, lang = "de", add = NULL, silent = F)
}
\arguments{
\item{ws_page}{url of one of the service pages contained in the menu under \url{http://ws-old.parlament.ch/}.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{add}{additional query modifiers.}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble. The number of columns depends on \code{ws_page}.
}
\description{
\code{get_overview} retrieves overview data from the parent entities of the Swiss Parliament WebServices.
}
\examples{
\donttest{
get_overview(ws_page = "http://ws-old.parlament.ch/affairs")
}
}
