% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview2}
\alias{get_overview2}
\title{Retrieve overview of all OpenParlData endpoints and fields}
\usage{
get_overview2(silent = F)
}
\arguments{
\item{silent}{if \code{TRUE}, no progress bar and messages are displayed.}
}
\value{
A tibble with the 2 columns \code{table} and \code{variable}.
}
\description{
\code{get_overview2} retrieves the names of the main resources provided by 
   the OpenParlData.ch REST API and the fields they contain.
}
\examples{
\dontrun{
# Overview for OpenParlData REST API
get_overview2()
}
}
