% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_parties_historic}
\alias{get_parties_historic}
\title{Retrieve historical party information}
\usage{
get_parties_historic(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{id} ID of party.
\item \code{updated} date of last update.
\item \code{abbreviation} short form of party.
\item \code{code} code of party.
\item \code{name} language-specific name of party.
}
}
\description{
\code{get_parties_historic} retrieves basic information on all parties in the history of the Swiss Parliament.
}
\examples{
\donttest{
get_parties_historic(lang = "fr")
}
}
