% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_related_tables2}
\alias{get_related_tables2}
\title{List related tables available}
\usage{
get_related_tables2(res)
}
\arguments{
\item{res}{an OpenParlData record (typically one row) as returned by
\code{\link{get_data2}()}.}
}
\value{
A sorted character vector containing the names of available related
  tables for the provided record.
}
\description{
\code{get_related_tables2} returns the names of related tables that
are available for an OpenParlData record.
}
\examples{
\dontrun{
# Retrieve one person
res <- get_data2("persons", firstname = "Gerhard", lastname = "Andrey")

# List available related resources for that record
get_related_tables2(res)
}
}
