% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_schedules}
\alias{get_schedules}
\title{Retrieve yearly schedules}
\usage{
get_schedules(year = 2010, lang = "de", silent = F)
}
\arguments{
\item{year}{specifies year. No data available for \code{year > 2015}.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 12 columns.
}
\description{
\code{get_schedules} retrieves yearly schedules of the Swiss Parliament.
}
\examples{
\donttest{
get_schedules(lang = "fr")
}
}
