% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_misc.R
\name{get_sessions}
\alias{get_sessions}
\title{Retrieve sessions}
\usage{
get_sessions(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{to} end of session.
\item \code{name} language-specific session name.
\item \code{from} start of session.
\item \code{code} code of session.
\item \code{updated} date of last update.
\item \code{id} session ID.
}
}
\description{
\code{get_sessions} retrieves basic information on all sessions of the Swiss Parliament since 1990.
}
\examples{
\donttest{
get_sessions(lang = "fr")
}
}
