% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Retrieve available tables}
\usage{
get_tables()
}
\value{
A character vector that contains all the names of the available tables.
}
\description{
\code{get_tables} retrieves the names of the available tables of the Swiss Parliament WebServices.
}
\examples{
\donttest{
# Get all available tables
get_tables()
}
}
