% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables2}
\alias{get_tables2}
\title{Retrieve available OpenParlData endpoints}
\usage{
get_tables2()
}
\value{
A sorted character vector containing the names of the available
  OpenParlData REST API endpoints.
}
\description{
\code{get_tables2} retrieves the names of the main REST API endpoints
provided by the OpenParlData.ch API.
}
\examples{
\dontrun{
# Get all available OpenParlData endpoints
get_tables2()
}
}
