% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Retrieve available variables}
\usage{
get_variables(table, pb.pos = NULL, pb = NULL)
}
\arguments{
\item{table}{name of the table to be searched. For an overview of available tables use \code{\link{get_tables}()}.}

\item{pb.pos}{value for the progress bar. Not to be specified outside of \code{\link{get_overview}()}.}

\item{pb}{progress bar. Not to be specified outside of \code{\link{get_overview}()}.}
}
\value{
A character vector that contains the names of the variables.
}
\description{
\code{get_variables} retrieves the variable names of a table of the Swiss Parliament WebServices.
}
\examples{
\dontrun{
# Get variables of table "Person"
get_variables(table = "Person")
}
}
