% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables2}
\alias{get_variables2}
\title{Retrieve available variables from an OpenParlData endpoint}
\usage{
get_variables2(table, pb.pos = NULL, pb = NULL)
}
\arguments{
\item{table}{name of the OpenParlData resource to be
queried. For an overview of available endpoints use
\code{\link{get_tables2}()}.}

\item{pb.pos}{value for the progress bar. Not to be
specified outside of \code{\link{get_overview}()}.}

\item{pb}{progress bar object. Not to be specified outside of
\code{\link{get_overview}()}.}
}
\value{
A sorted character vector containing the names of the fields
  available in the selected OpenParlData endpoint.
}
\description{
\code{get_variables2} retrieves the field names of a resource provided by
the OpenParlData.ch REST API.
}
\examples{
\dontrun{
# Get variables of OpenParlData resource "persons"
get_variables2(table = "persons")
}
}
