% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_votes_affairs2}
\alias{get_votes_affairs2}
\title{Retrieve all votes on affair level}
\usage{
get_votes_affairs2(lang = "de", silent = F)
}
\arguments{
\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 17 columns, including the voting results.
}
\description{
\code{get_votes_affairs2} is wrapper around \code{get_votes_affairs} and enables the mass retrieval of
    vote data of the National Council provided by the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_votes_affairs2()
}
}
