% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_votes_councillors}
\alias{get_votes_councillors}
\title{Retrieve individual vote decisions}
\usage{
get_votes_councillors(councillor_id, from = NULL, to = NULL,
  legislative_period = NULL, session = NULL, search_text = NULL,
  affair_id = NULL, lang = "de", silent = F)
}
\arguments{
\item{councillor_id}{numeric or character vector of councillor IDs to query.}

\item{from}{specifies start of observation period, filters votes by date. Recommended date format \code{YYYY/MM/DD}.}

\item{to}{specifies end of observation period, filters votes by date. Recommended date format \code{YYYY/MM/DD}.}

\item{legislative_period}{specifies observation period by legislative period.  See \code{get_legislative_periods} for IDs.}

\item{session}{specifies observation period by session. See \code{get_sessions} for IDs.}

\item{search_text}{character string, subsets votes by search text.}

\item{affair_id}{affair ID. See \code{get_affairs} for IDs.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 16 columns, including \code{councillorVote.decision}, \code{meaningYes}, \code{meaningNo} and \code{divisionText}.
}
\description{
\code{get_votes_councillors} retrieves individual vote decisions of members of the National Council provided by the
    WebServices of the Swiss Parliament.
}
\examples{
\donttest{
# All votes of Thomas Aeschi and Cédric Wermuth on USRIII
get_votes_councillors(councillor_id = c(2758, 2759), affair_id = 20150049)

# All votes of Pascale Bruderer Wyss during her time in the National Council
get_votes_councillors(councillor_id = 2573)
}
}
