% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissparl.R
\docType{package}
\name{swissparl-package}
\alias{swissparl}
\alias{swissparl-package}
\title{swissparl: The Swiss Parliament Webservices R API}
\description{
The Swiss Parliament Webservices R API.
}
\details{
See the README on
\href{https://github.com/zumbov2/swissparl#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.parlament.ch/en/services/open-data-webservices}
  \item \url{https://api.openparldata.ch}
  \item Report bugs at \url{https://github.com/zumbov2/swissparl/issues}
}

}
\author{
\strong{Maintainer}: David Zumbach \email{david.zumbach@gfzb.ch}

Other contributors:
\itemize{
  \item Benjamin Gföhler [contributor]
}

}
\keyword{internal}
