% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissparl.R
\docType{package}
\name{swissparl}
\alias{swissparl}
\alias{swissparl-package}
\title{\code{swissparl} package}
\description{
The Swiss Parliament Webservices R API
}
\details{
See the README on
\href{https://github.com/zumbov2/swissparl#readme}{GitHub}
}
