% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_councillors.R
\name{get_councillors}
\alias{get_councillors}
\title{Retrieve councillor details}
\usage{
get_councillors(councillor_id, lang = "de", silent = F)
}
\arguments{
\item{councillor_id}{numeric or character vector of councillor IDs to query.}

\item{lang}{specifies language. Available are German ("de"), French ("fr"), Italian ("it") and English ("en").}

\item{silent}{if \code{TRUE}, no progress bar is displayed.}
}
\value{
A tibble with 40 columns, including \code{canton}, \code{birthDate}, \code{committeeMemberships},
    \code{maritalStatus}, \code{numberOfChildren} and \code{militaryGrade}.
}
\description{
\code{get_councillors} retrieves detailed data on councillors provided by the WebServices of the Swiss Parliament.
}
\examples{
\donttest{
get_councillors(councillor_id = c(823, 316))
}
}
