% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{Retrieve overview of all tables and variables}
\usage{
get_overview(silent = F)
}
\arguments{
\item{silent}{if \code{TRUE}, no progress bar and messages are displayed.}
}
\value{
A tibble with the 2 columns \code{table} and \code{variable}.
}
\description{
\code{get_overview} retrieves the names of all available tables of the Swiss Parliament WebServices and
    the variables they contain.
}
\examples{
\dontrun{
get_overview()
}
}
